+++ src/Domain/DiagramaConverter.java
@@ -35,7 +35,7 @@ public class DiagramaConverter implements EntitateConverter<Diagrama>{
         while(index < tokens.length)
         {
             //trecem in for cate 3,pentru a lua fiecare obiect de tip cheltuiala
-            Cheltuiala c = new Cheltuiala(Integer.parseInt(tokens[index]) , tokens[index+1] , Integer.parseInt(tokens[index+2]));
+            Cheltuiala c = new Cheltuiala(Integer.parseInt(tokens[index]) , tokens[index+1] , Float.parseFloat(tokens[index+2]));
             
             lista_cheltuieli.add(c);
             index=index+3;//pt a incepe de la urmatorul obiect Cheltuiala
